requirements.txt
pexpect==4.3.1
pika==0.11.2
PyYAML==3.12
